+++ test/fixtures/laravel-api/routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\TokenAuthenticationController;
 use App\Http\Resources\UserResource;
 use Illuminate\Foundation\Http\Middleware\HandlePrecognitiveRequests;
 use Illuminate\Http\Request;
@@ -8,6 +9,12 @@ use Illuminate\Support\Facades\Route;
 Route::get('/user', function (Request $request) {
     return UserResource::make($request->user());
 })->middleware('auth:sanctum');
+
+Route::post('/tokens/create', [TokenAuthenticationController::class, 'store'])
+->middleware(['guest']);
+
+Route::post('/tokens/delete', [TokenAuthenticationController::class, 'destroy'])
+->middleware(['auth:sanctum']);
 
 Route::patch('/profile', function (Request $request) {
     $request->validate([